Fixed user functions, maybe fixed aliases and list functions

+++ cmder/config/user_profile.ps1
@@ -3,7 +3,9 @@ Copy-Item "$env:userprofile\Desktop\Script\script.ps1" `
 "$env:userprofile\Documents\WindowsPowershell\Microsoft.PowerShell_profile.ps1"
 
 # Delete unecessary git folder
-Remove-Item "$env:userprofile\Desktop\Script\.git" -Recurse -Force
+if ((Test-Path "$userprofile\Desktop\Script\.git") -eq $true) {
+    Remove-Item "$env:userprofile\Desktop\Script\.git" -Recurse -Force
+}
 
 # Run script
 . $profile
